beautify minor things, make compilation more strict (thanks Sven for hints)

+++ functions.c
@@ -27,12 +27,12 @@ double time_diff( struct timeval *start, struct timeval *end ) {
 	unsigned long sec = (unsigned long) end->tv_sec - start->tv_sec;
 	unsigned long usec = (unsigned long)end->tv_usec - start->tv_usec;
 
-	if(sec > end->tv_sec) {
+	if(sec > (unsigned long)end->tv_sec) {
 		sec += 1000000000UL;
 		--sec;
 	}
 
-	if(usec>end->tv_usec) {
+	if(usec > (unsigned long)end->tv_usec) {
 		usec += 1000000000UL;
 		--usec;
 	}
